Add map support to realist
