+++ src/frontend/realist-app/main/components/map_components/leafletMap.css
@@ -1,0 +1,3 @@
+.leaflet-container { 
+    height: 75vh;
+}
