+++ src/frontend/realist-app/main/components/table_components/TruliaDataTable/TruliaDataTable.tsx
@@ -6,7 +6,7 @@ import "ag-grid-community/styles/ag-theme-quartz.css"; // Optional Theme applied
 import { Box, Flex, Text } from "@chakra-ui/react";
 import { ColDef } from "ag-grid-community";
 import "../agGridTable.css"
-import { formatDbDate, formatNumberToMoney } from "../../../helpers/common";
+import { formatNumberToMoney, returnTruliaFullAndSummaryListFromApiTruliaResponse } from "../../../helpers/common";
 
 interface TruliaDataTableProps {
     isFetching: boolean
@@ -37,90 +37,11 @@ export function TruliaDataTable ({isFetching, onListingClick, listings}: TruliaD
     
     useEffect(()=>{
         if (listings){
-            let listingsSummaryList: TruliaListingSummary[] = []
-            let listingsFullList: TruliaListingFull[] = []
-            for (let listing of listings){
-                let listingsFull = convertApiResponseToTruliaListingObject(listing);
-                listingsFullList.push(listingsFull)
-                let listingsSummary = summarizeListing(listingsFull);
-                listingsSummaryList.push(listingsSummary)
-            }
-            setFullRowsData(listingsFullList)
-            setRowsData(listingsSummaryList)
+            let listingsList = returnTruliaFullAndSummaryListFromApiTruliaResponse(listings)
+            setFullRowsData(listingsList.listingsFullList)
+            setRowsData(listingsList.listingsSummaryList)
         }
     }, [listings])
-
-    const convertApiResponseToTruliaListingObject = (apiListingObject: ApiTruliaListingResponse) => {
-        let listingObjectPriceHistoryList: PriceHistory[] = [];
-        if (apiListingObject.price_history){
-            for (let apiListingObjectPriceHistory of apiListingObject.price_history) {
-                let listingObjectPriceHistory: PriceHistory = {
-                    event: apiListingObjectPriceHistory.event,
-                    formattedDate: apiListingObjectPriceHistory.formattedDate,
-                    formattedPrice: apiListingObjectPriceHistory.price?.formattedPrice,
-                    priceChange: apiListingObjectPriceHistory.priceChange ? {
-                        formattedPriceChangePercent: apiListingObjectPriceHistory.priceChange.formattedPriceChangePercent,
-                        priceChangeDirection: apiListingObjectPriceHistory.priceChange.priceChangeDirection,
-                        priceChangePercent: apiListingObjectPriceHistory.priceChange.priceChangePercent,
-                        priceChangeValue: apiListingObjectPriceHistory.priceChange.priceChangeValue?.formattedPrice
-                    } : undefined
-                }
-                listingObjectPriceHistoryList.push(listingObjectPriceHistory);
-            }
-        }
-        
-        let listingFull: TruliaListingFull = {
-            key: apiListingObject.key,
-            address: apiListingObject.address,
-            askingPrice: apiListingObject.asking_price,
-            bedrooms: apiListingObject.bedrooms,
-            bathrooms: apiListingObject.bathrooms,
-            neighborhood: apiListingObject.neighborhood,
-            city: apiListingObject.city,
-            zip: apiListingObject.zip,
-            floorSqft: apiListingObject.floor_sqft,
-            propertyType: apiListingObject.property_type,
-            propertySubType: apiListingObject.property_subtype,
-            yearBuilt: apiListingObject.year_built,
-            yearRenovated: apiListingObject.year_renovated,
-            dateListedOrSold: apiListingObject.date_listed_or_sold ? formatDbDate(apiListingObject.date_listed_or_sold) : null,
-            dateScraped: formatDbDate(apiListingObject.date_scraped),
-            architecture: apiListingObject.architecture,
-            basement: apiListingObject.basement,
-            condition: apiListingObject.condition,
-            foundation: apiListingObject.foundation,
-            houseMaterial: apiListingObject.house_material,
-            structureType: apiListingObject.structure_type,
-            lotSqft: apiListingObject.lot_sqft,
-            parking: apiListingObject.parking,
-            trulia_url: apiListingObject.trulia_url,
-            priceHistory: listingObjectPriceHistoryList,
-            mlsListingId: apiListingObject.mls_listing_id,
-            description: apiListingObject.description,
-        }
-
-        return listingFull;
-    }
-
-    const summarizeListing = (listing: TruliaListingFull) => {
-        let listingSummary: TruliaListingSummary = {
-            key: listing.key,
-            address: listing.address,
-            askingPrice: listing.askingPrice,
-            bedrooms: listing.bedrooms,
-            bathrooms: listing.bathrooms,
-            neighborhood: listing.neighborhood,
-            floorSqft: listing.floorSqft,
-            propertyType: listing.propertyType,
-            yearBuilt: listing.yearBuilt,
-            dateListedOrSold: listing.dateListedOrSold,
-            dateScraped: listing.dateScraped,
-            description: listing.description
-        }
-        return listingSummary;
-    }
-
-
     
     return (
         <Flex width="100%" height="100%" minHeight="250px" className="ag-theme-quartz">
