+++ src/frontend/realist-app/main/helpers/globalInterfaces.tsx
@@ -27,8 +27,10 @@ export interface ApiTruliaListingResponse {
     foundation: string | null;
     house_material: string | null;
     key: string;
+    latitude: number | null;
     listing_status: string | null;
     location: string | null;
+    longitude: number | null;
     lot_sqft: number | null;
     mls_listing_id: string | null;
     neighborhood: string | null; 
@@ -108,6 +110,8 @@ export interface TruliaListingFull{
     houseMaterial: string | null;
     structureType: string | null;
     lotSqft: number | null;
+    latitude: number | null;
+    longitude: number | null;
     parking: string | null;
     trulia_url: string | null;
     priceHistory: PriceHistory[] | null;
